PR #3465 Made the required changes
